+++ src/main/java/model/dao/StudentDAO.java
@@ -22,7 +22,7 @@ public class StudentDAO {
 	}
 	
 	// user_id(FK)을 통해서 학생테이블을 가져옴
-	public StudentsDTO findStudentByUserId(int userId) {
+	public StudentsDTO findStudentByUserId(long userId) {
 		StudentsDTO student = new StudentsDTO();
 		
 		String sql = "SELECT * FROM student WHERE user_id = ?";
@@ -30,7 +30,7 @@ public class StudentDAO {
 		try (Connection conn = DBConnection.getConnection())
 		{
 			PreparedStatement pstmt = conn.prepareStatement(sql);
-			pstmt.setInt(1, userId);
+			pstmt.setLong(1, userId);
 			
 			ResultSet rs = pstmt.executeQuery();
